updated max func

+++ func.cc
@@ -54,7 +54,7 @@ int mean(int a,int b, int c){
   return res;
 }
 int max(double a[],int size){//Daniil Novikov
-  double max = a[0];
+  int max = a[0];
   for(int i=1;i<size;i++){
     if(a[size]>max){
       max = a[size];
